HTTP output plugin optimizations.
Don't copy the input buffer, just use it as is. Keep the mutex locked throughout
main loop. This will ensure the input and output threads will have exclusive
access to the buffer. If you want to allow for the input buffer to continue
reading while the buffer is being read, use a double buffer approach.